refactor(plan-detail): move usePlanDetailPage and types into shell/hooks/
