+++ frontend/src/react/pages/project/plan-detail/components/PlanDetailApprovalFlow.test.tsx
@@ -6,7 +6,7 @@ import { ApprovalStatus } from "@/types/proto-es/v1/common_pb";
 import type { Issue } from "@/types/proto-es/v1/issue_service_pb";
 import { Issue_Approver_Status } from "@/types/proto-es/v1/issue_service_pb";
 import { PlanDetailProvider } from "../context/PlanDetailContext";
-import type { PlanDetailPageState } from "../hooks/usePlanDetailPage";
+import type { PlanDetailPageState } from "../shell/hooks/types";
 import { PlanDetailReviewApprovalFlow } from "./PlanDetailApprovalFlow";
 
 (
