+++ frontend/src/react/pages/project/plan-detail/components/PlanDetailChangesBranch.test.tsx
@@ -6,7 +6,7 @@ import { State } from "@/types/proto-es/v1/common_pb";
 import type { PlanCheckRun } from "@/types/proto-es/v1/plan_service_pb";
 import type { CheckReleaseResponse_CheckResult } from "@/types/proto-es/v1/release_service_pb";
 import { PlanDetailProvider } from "../context/PlanDetailContext";
-import type { PlanDetailPageState } from "../hooks/usePlanDetailPage";
+import type { PlanDetailPageState } from "../shell/hooks/types";
 import { PlanDetailChangesBranch } from "./PlanDetailChangesBranch";
 
 (
