+++ frontend/src/react/pages/project/plan-detail/context/PlanDetailContext.tsx
@@ -1,5 +1,5 @@
 import { createContext, type ReactNode, useContext } from "react";
-import type { PlanDetailPageState } from "../hooks/usePlanDetailPage";
+import type { PlanDetailPageState } from "../shell/hooks/types";
 
 const PlanDetailContext = createContext<PlanDetailPageState | null>(null);
